+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased](https://github.com/kabamgamer/dd2-strategy-hub/tree/main) (yyyy-mm-dd)
 
 * Defense calculators: Prevents the user from adding duplicate shards
+* Community maps: Added "Requires 2 accounts" tag (with that tag you can make fair use of 4 mass destructions)
+* 
 
 [Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.7.1...main)